[SDSAET-50] Rename Metric to DistanceFunction
