+++ core/jobs/src/main/java/com/cognifide/aet/job/common/groupers/DistanceFunction.java
@@ -4,7 +4,7 @@
  * Copyright (C) 2013 Cognifide Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License. You may obtain a copy of the License at
+ * in compliance with the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -13,9 +13,11 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.job.common.groupers.algorithm;
+package com.cognifide.aet.job.common.groupers;
 
-public interface Metric<T> {
+import java.util.function.BiFunction;
 
-  double calculate(T a, T b);
+// todo javadoc
+public interface DistanceFunction<T> extends BiFunction<T, T, Double> {
+
 }
