+++ core/jobs/src/main/java/com/cognifide/aet/job/common/groupers/jserrors/JsErrorsDistanceFunction.java
@@ -4,7 +4,7 @@
  * Copyright (C) 2013 Cognifide Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License. You may obtain a copy of the License at
+ * in compliance with the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -13,17 +13,22 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.job.common.groupers.algorithm;
+package com.cognifide.aet.job.common.groupers.jserrors;
 
 import com.cognifide.aet.job.api.collector.JsErrorLog;
+import com.cognifide.aet.job.common.groupers.DistanceFunction;
 import org.apache.commons.text.similarity.LevenshteinDistance;
 
-public class JsErrorMetric implements Metric<JsErrorLog> {
+public class JsErrorsDistanceFunction implements DistanceFunction<JsErrorLog> {
+
+  private final LevenshteinDistance levenshteinDistance = new LevenshteinDistance();
 
   @Override
-  public double calculate(JsErrorLog a, JsErrorLog b) {
-    int distance = new LevenshteinDistance().apply(a.getErrorMessage(), b.getErrorMessage());
-    int longer = Math.max(a.getErrorMessage().length(), b.getErrorMessage().length());
+  public Double apply(JsErrorLog a, JsErrorLog b) {
+    String s1 = a.getErrorMessage();
+    String s2 = b.getErrorMessage();
+    int distance = levenshteinDistance.apply(s1, s2);
+    int longer = Math.max(s1.length(), s2.length());
     return 1 - ((longer - distance) / (double) longer);
   }
 }
